Fix misinterpreted EF in log-N deviate formulas
The two-sided formulas
were a wrong approach for computing the distribution with error factors.
The formulas are corrected
to work with one-sided, lower-tailed bounds,
as provided in the MEF standard.

As a result of the correction,
the sample standard deviations of the Monte-Carlo simulations increased,
and the simulated probabilities became closer to the mean values.

In addition, the maximum value evaluation now computes 99.9% percentile
instead of 99% (just for the sake of simplicity of 3 sigmas).

Fixes #147