Use CMAKE_CXX_STANDARD to set compiler C++14 flags
On CMake 3.1 or later this feature is available.
Unfortunately, this feature is not implemented for Intel compilers,
so the C++ version flag is still set manually for this compiler.
If cmake is older,
manual checking and setting are performed as usual.
In addition, minimum compiler versions are checked in CMake setup.

Qt 5.7 or later seems to require this flag to properly build with C++14.