Redesign unique index assignment in PDAG
The Node index assignment is delegated to the PDAG.
The graph as a whole ensures
that each new node gets an appropriate index.
Moreover, upon construction from a Fault Tree,
it guarantees that Basic Events are converted before Gates,
ensuring sequential low indices for Variables.
This allows for efficient (index->data) mapping
for later analysis classes.

This approach eliminates global static variables
that used to manage the unique indices for Node class instances.