Introduce Pdag::IndexMap<T> for index->data map
This container is based on ext::index_map non-zero-based index vector.
The container adaptor type encapsulates
the efficient mapping of variable indices to some arbitrary data.

Other analysis facilities do not need to work with a raw vector
and shift the indices back and forth.

However, ext::index_map is limited to this kind of use cases only.
It is not a general purpose adapter
but rather operator[] adjuster.