Make Gate::args<T>() const correct
No shared pointer is leaked or reference counts are manipulated
upon call from const overload of the function.
It returns the arguments by reference instead of shared pointers
to prevent the modifications.
In addition, it enforces transitive constness
by returning const references of arguments
so that the Boolean formula cannot be changed indirectly.