Implement a set iterator for ZBDD with modules
This is very drastic rewrite of the iterator for ZBDD w/o modules.
Dealing with modules non-recursively complicates the matters a lot.
The recursion is replaced with three stacks:
product, node, module;
Each module needs its own iterator
due to being a separate ZBDD altogether.

This initial implementation is functional,
yet it is not optimal or elegant.
However, this iterator completely replaces
the previous product generation implementations and approaches.
Moreover, the ZBDD is no longer destroyed by the analysis
and product extraction.

Issue #159