Replace result vector<vector<int>> with Zbdd
Analysis facilities work directly with the resultant ZBDD
without converting it to other data structures.
The performance gain is noticeable
only for a large of number of products (1 Million or more).
The analysis speedup ranges from 10% to 5x;
memory consumption drops by 1.5-2x;

The downside is that iteration over ZBDD products is 2x slower
than iteration over a vector of products.

Issue #159