Move MissionTime from Constant to Parameter
For probability analysis with different mission times,
it needs to act more like parameter than a constant.
The mission time is a very special lightweight parameter.

To allow changes in mission time,
a setter function provided for its time value.

The Units enum is moved from Expression to Parameter
because it is only used in Parameter.

Issue #161