Implement probability evaluation over time period
The probability analysis is made aware
of the MissionTime expression of a model under analysis.
The evaluation takes place when the time-step is non-zero.

The p vs. time is reported with the MEF "curve" feature;
however, the standard schema is a bit strict with units,
so the current report doesn't yet validate.

Issue #161