Set Expression::MissionTime::Min() to 0
For validation and probability analysis purposes,
the mission time can vary from 0 to the user specified amount.
It is infeasible to validate for all time points in [0, mission_time].
Instead, the minimum and maximum time
is used as conservatively as possible.

The Weibull Expression is adjusted to work
when the time-shift is less than the mission time.