Implement simplistic SIL PFH calculation
The approach is very simplistic by using the current unavailability
instead of unconditional failure intensity.
From couple of benchmarks,
it has been observed to result in close values of the same magnitude.

In any case, PFH over a fault tree is inherently approximate,
and this approach simplifies the implementation noticeably.

Closes #160