Make CCF factor levels optional
CCF factor levels can be deduced
from CCF models and the order of factors in the input.
The current implementation follows the same strategy as C/C++ Enums;
that is, the next level is previous level + 1.
The starting level depends on a specific CCF model.
There can also be gaps and mixing with explicitly defined levels.