Refactor Event lifetime management
Model, fault tree, and other container own Events.
In gates (formulas), only references to these Events are given
without the need to care about their lifetime.
It is guaranteed to be as long as the lifetime of the Model.

CCF groups also create their own CCF Events and keep them alive
while the Model is alive.

This approach removes unnecessary lifetime management by the formulas.