Remove implicit handling of NULL/pass-through gate
Instead of implicitly resetting the formula parent for the NULL gate,
the single argument is directly added to the null gate explicitly.

The implicit handling has lead to errors due the parent XML node
is not a formula and has label/attributes,
which show up in XPATH query in formula argument iterations.

Fixes #166