Refactor cycle to produce container of pointers
Instead of assuming that nodes have 'name' member function,
the node addresses are used as identifiers.
Only upon printing,
nodes are required to have 'id' function that results in string.
The reason for the id function is that the nodes may be private.
In other words, this change also fixes the bug
of misreporting nodes with public names with possible name clashes.