Implement setting house events in event-tree walks
The house-event values can be changed in path collection phases
as described in the MEF.
This feature and other set-instructions
truly require deep copy of fault trees.
The current implementation is naive, not optimal.
It doesn't memoize or avoid unnecessary copies.
These optimization are expected to be implemented
after major features (other set-instructions and directions)
are in place.

Issue #179
Closes #179