Implement test-event context in event-tree walk
Upon event-tree walk,
a context of functional and initiating events is gathered
to help test-{functional,initiating}-event expressions to evaluate.
This context is per model,
so no two event-trees can be walked concurrently.
This choice is made due to the current implementation of expressions
without any visitors.
No other expressions need a visitor for this purpose
except for test-event expressions (2 of them),
so rewriting the whole expression implementation is not practical.

Issue #68
Closes #68