GUI: Keep find/filter string unique per tab
The initial idea of a global search via find/filter is confusing
because it misleads that the next tab is filtered as well.
Instead, keep find/filter strings per tab for clarity.

The more advanced find/filter needs to be implemented
for global searches with history, autocompletion,
and options like case-sensitivity, pattern, regex, whole-word, etc.