Implement owner_ptr (unique_ptr + ownership bit)
Instead of relying on nullptr to signify ownership,
owner_ptr uses explicit flag.
This allows preservation of the original pointer after move
unlike unique_ptr, which resets its pointer to nullptr.
However, dereferencing after move is unsafe but not undefined.
If the user of moved-from owner_ptr knows
that the pointed object is not expired,
dereferencing succeeds.

The preserve-after-move functionality
is the main reason for this pointer.
This pointer is necessary
to emulate 'extract' function in boost::multi_index.