GUI: Add 'container' box to EventDialog
This is the destination container for event definitions.
Currently, only gates are defined within fault trees.
Other event types are defined in the Model directly.

The container section can be expanded in future
to choose the destination files as well upon working with multi-files.

The current new-gate definitions
are constrained to define a fault-tree as well,
so that only single-top fault tree are allowed.

Issue #204
Issue #209