GUI: Install crash handlers for internal errors
This is a basic and best-effort approach
based on intercepting the system interrupt and terminate signals.
It may not work with multi-threaded code
(undefined behavior and Qt event-filter violation),
and it may not be cross-platform (Windows, in particular).

These errors are best effort to guard users from crashes
due to internal application bugs.
The bugs are not ignored, but reported with a dialog,
which may also crash itself (e.g., std::bad_alloc).