GUI: Check for cycles upon gate arg modification
In EventDialog,
if an arg would introduce a cycle,
it is not added to the gate args list.
An explanation message is printed in the dialog.

The current implementation for cycle detection is not optimal.
It is exponential and based on descendants
rather than ancestors.
This is due to current limitations of the core API for mef::Event.

Issue #133