Remove ext::owner_ptr
The only reason for the owner_ptr was
to extract unique_ptr out of boost::multi_index_container.
There's another way of doing it w/o owner_ptr.
The extract function is refactored with const_cast and move
of the original unique_ptr.
This is still ugly but seems to work.