Move role-based lookup from Model to Initializer
This is an initialization issue from XML MEF format.
It is the concern of the initializer.
The original reason
for the implementation of this functionality in Model
was the performance and access to inner containers of the Model.
This inner containers are moved from Model to Initializer
to keep the concerns separate.
The Model should be agnostic how it is being initialized.