GUI: Make diagram width units relative to height
Instead of using independent width and height,
the width units are fixed to be a half of the height units.
This is the original assumption of the diagram dimensions.
The height unit is still derived from the height metrics of the font
to make the graphics screen DPI independent.
With these changes, the diagram graphics size units
have only single degree of freedom,
which should minimize
the differing looks on different operation systems.

Issue #215