GUI: Configure bundling w/ shared libs for Windows
The shared libraries are preferred with mingw64.
This approach is less failure prone than static builds on Windows.
Moreover, it allows using other Qt-based libraries like Qwt in future.

The noticeable negative effect is that the bundle size increases 2x.

Another issue is with Wine failing to launch qwindows.dll
for some reason. This needs to be fixed in future.

Issue #217