Implement ExternFunction initialization from XML
The implementation follows
the MEF proposal for "define-extern-function".
This allows any extern C function to be declared in the MEF XML
and used later in Expression.

The downside is exponential explosion of possible function-interfaces
with respect to allowed parameter types and numbers.
In particular, the implementation relies on
C++ template meta-programming,
which heavily slows the build times
and increases build memory consumption.

The binary size does not increase much with the current implementation.

Validation:
    - Duplicate declarations
    - Empty return type
    - Invalid number of parameters exceeding the maximum number
    - Undefined extern library to lookup the symbol
    - Undefined extern-function symbol in the reference library

Issue #74