Make MEF extern library & function features opt-in
Extern library and function features allow
execution of arbitrary code from arbitrary libraries.
This is unsafe
and should be allowed only for trusted input and libraries.
By default, the dynamic library loading with XML is disabled.
This can only be enabled at command-line (``--allow-extern``) or API.
No XML input should be able to subvert and enable this feature.

Issue #74