Optimize xml::Stream output indentation
Instead of using std::string based indentation,
provide static buffer per document
to produce static indentation string on the stack.

This optimization is possible only
due to the nature of the XML streaming.
That is, indentation string is used only by a single element at a time.
The Stream document contains the string buffer,
which is provided to the element being streamed.

This improves performance, but unfortunately,
std::ostream is too slow to see the difference
in actual analysis report printing.

The tests are reworked to be fixtures
because now XML Elements cannot be created without Stream document.