Convert stdio errors to IO exception in XML Stream
The IO errors should not be ignored
when file integrity is of concern.
Serialization and reporting are now IO write-failure aware.

For efficiency reasons, the stdio FILE error code is tested
only at the end of XML Streaming in the RAII class destructor.
The IO exceptions are thrown conservatively
only if there are no other exceptions on flight.
More detailed stack unwinding w/ unhandled exception
can be detected in future with C++17 'uncaught_exceptions' function.

Issue #228