GUI: Implement fail-safe IO write w/ boost::rename
This is not the cleanest solution.
The temporary file is created in the same file system directory,
and then rename/move operation is performed.
This approach assumes that user has permission
to create files in the directory.
Moreover, it may mess up the file permission of the original file.

The rename operation will not succeed cleanly
if the original file is a symlink.
In this case, the symlink is replaced with a new file.

Issue #228