Apply declarative substitutions in PDAG ctor
Declarative substitutions (e.g., delete terms, recovery rules)
are applied to the Boolean formula directly (=> operator)
as described in the Open-PSA MEF.

PDAG needs to be aware of all the events in the substitutions
because they may not appear in the original fault tree for construction.
The Model is passed directly to PDAG to gather substitution events.
This is convenient but not the ideal approach.
It would be more appropriate to pass
only the relevant set of substitutions instead of the whole model.

Issue #154