Apply non-declarative Substitutions in ZBDD
The implementation is naive with O(N * M) complexity.
It doesn't take advantage of ZBDD features.
The reason for suboptimal implementation is
that non-declarative substitutions should be avoided in the first place.

This will be optimized if there is more need for it in future.
In addition, more validation/verification tests are required
before taking advantage of ZBDD features.

The current solution for Substitution application is adhoc
because all/most facilities assume declarative constructs,
and non-declarative approaches are bolted-in.

Issue #154