GUI: Discover languages at runtime in main()
Instead of hard-coding available languages,
the main() tries loading the preference or system language
from the known translation directories.
If it doesn't succeed, it defaults to English (untranslated).

This makes enabling or addition of new languages easy.
It is decoupled from the compilation of SCRAM
and can be treated as a separate package altogether.
To add a new language,
just install the translations into the SCRAM translation directory.

Issue #242