GUI: Add lifetime and address stability contracts [skip ci]
This is required for an efficient implementation of undo-redo commands.
Objects must have stable addresses
to be referenced w/ pointers/references on the undo stack.
Moreover, an object must be alive as long as
there's a command in the stack with a reference to the object.

This may sound like a use case for a shared pointer,
but it is not necessary
thanks to the linearity of the undo stack (i.e., timeline/history).
The object lifetime is determined automatically
by the lifetime of the constructive/destructor commands
similar to the object's constructor/destructor;
however, in this case, the undo stack plays the role of program stack.
This allows a safe use of raw pointers or references
instead of shared pointers (or a garbage collector)
as long as all the undo-redo (state changing) commands
are registered with the undo stack.