GUI: Detect modelReset for removable tables
It turns out in Qt,
current selection and item signals are not emitted
when the model is reset.

The model reset signal needs to be connected
to selection dependent actions separately.

This fixes the issue with Gate table having "Remove" active
with no selections available.
However, with Gate tables, there's much involved problem
with resetting the model in the first place
(can't track formula changes precisely).

Fixes #241