Implement vector-based ext::linear_set
The vector-based set is a convenience adaptor
when the insertion order must be preserved with unique values.
This is only efficient for a small number of values
since the algorithmic complexity is linear O(N).

In addition, it acts like map
(akin to boost::multi_index and Python set)
with key extractor policy.