Revert to bool cmd-line options as implicit false
This used to be for convenience
to override config file options.
However, the inconvenience is that
at command-line, bool flags need an argument.
Moreover, it is inconsistent even within the bool flags.
Instead, revert to a simpler setup
with bool flags having no arguments.
The command-line flags still override the config flags
but only with the enabling (non-suppressive) effect.

Issue #266